Rename `topology::kinds` to `objects`
